[googlemaps] Extract drawing.d.ts
