+++ types/googlemaps/tsconfig.json
@@ -22,6 +22,7 @@
         "reference/coordinates.d.ts",
         "reference/event.d.ts",
         "reference/control.d.ts",
+        "reference/drawing.d.ts",
         "reference/street-view-service.d.ts",
         "reference/places-widget.d.ts",
         "googlemaps-tests.ts"
